Agile-Office-Main: refactor main screen && Button component typescript && dark mode fix && navigation screen theme and import fix
